+++ include/extractor/guidance/intersection.hpp
@@ -7,7 +7,6 @@
 
 #include "extractor/guidance/turn_instruction.hpp"
 #include "util/bearing.hpp"
-#include "util/guidance/toolkit.hpp"
 #include "util/node_based_graph.hpp"
 #include "util/typedefs.hpp" // EdgeID
 
@@ -29,16 +28,15 @@ struct IntersectionShapeData
 inline auto makeCompareShapeDataByBearing(const double base_bearing)
 {
     return [base_bearing](const auto &lhs, const auto &rhs) {
-        return util::bearing::angleBetweenBearings(base_bearing, lhs.bearing) <
-               util::bearing::angleBetweenBearings(base_bearing, rhs.bearing);
+        return util::angleBetweenBearings(base_bearing, lhs.bearing) <
+               util::angleBetweenBearings(base_bearing, rhs.bearing);
     };
 }
 
 inline auto makeCompareAngularDeviation(const double angle)
 {
     return [angle](const auto &lhs, const auto &rhs) {
-        return util::guidance::angularDeviation(lhs.angle, angle) <
-               util::guidance::angularDeviation(rhs.angle, angle);
+        return util::angularDeviation(lhs.angle, angle) < util::angularDeviation(rhs.angle, angle);
     };
 }
