+++ include/extractor/guidance/turn_lane_data.hpp
@@ -4,7 +4,6 @@
 #include "extractor/guidance/turn_lane_types.hpp"
 #include "util/attributes.hpp"
 #include "util/typedefs.hpp"
-#include <string>
 #include <vector>
 
 namespace osrm
