+++ include/extractor/guidance/turn_lane_types.hpp
@@ -8,11 +8,9 @@
 #include <unordered_map>
 #include <vector>
 
-#include <boost/assert.hpp>
 #include <boost/functional/hash.hpp>
 
 #include "util/json_container.hpp"
-#include "util/log.hpp"
 #include "util/typedefs.hpp"
 
 namespace osrm
