+++ src/extractor/extractor.cpp
@@ -10,7 +10,6 @@
 
 #include "extractor/raster_source.hpp"
 #include "storage/io.hpp"
-#include "storage/io.hpp"
 #include "util/exception.hpp"
 #include "util/exception_utils.hpp"
 #include "util/graph_loader.hpp"
@@ -44,7 +43,6 @@
 #include <algorithm>
 #include <atomic>
 #include <bitset>
-#include <chrono>
 #include <chrono>
 #include <fstream>
 #include <iostream>
