+++ src/extractor/guidance/intersection.cpp
@@ -1,5 +1,4 @@
 #include "extractor/guidance/intersection.hpp"
-#include "extractor/guidance/toolkit.hpp"
 
 #include <boost/range/adaptor/transformed.hpp>
 #include <boost/range/algorithm/find_if.hpp>
@@ -9,6 +8,8 @@
 #include <algorithm>
 #include <functional>
 #include <limits>
+
+using osrm::util::angularDeviation;
 
 namespace osrm
 {
