+++ src/extractor/guidance/motorway_handler.cpp
@@ -1,18 +1,17 @@
 #include "extractor/guidance/motorway_handler.hpp"
 #include "extractor/guidance/constants.hpp"
 #include "extractor/guidance/road_classification.hpp"
-#include "extractor/guidance/toolkit.hpp"
 
-#include "util/guidance/toolkit.hpp"
-#include "util/log.hpp"
+#include "util/bearing.hpp"
+#include "util/guidance/name_announcements.hpp"
 
 #include <limits>
 #include <utility>
 
 #include <boost/assert.hpp>
 
-using osrm::util::guidance::angularDeviation;
-using osrm::util::guidance::getTurnDirection;
+using osrm::util::angularDeviation;
+using osrm::extractor::guidance::getTurnDirection;
 
 namespace osrm
 {
@@ -197,9 +196,6 @@ Intersection MotorwayHandler::fromMotorway(const EdgeID via_eid, Intersection in
         else if (countValid(intersection) > 0) // check whether turns exist at all
         {
             // FALLBACK, this should hopefully never be reached
-            util::Log(logDEBUG) << "Fallback reached from motorway, no continue angle, "
-                                << intersection.size() << " roads, " << countValid(intersection)
-                                << " valid ones.";
             return fallback(std::move(intersection));
         }
     }
@@ -275,7 +271,6 @@ Intersection MotorwayHandler::fromMotorway(const EdgeID via_eid, Intersection in
                                              via_eid,
                                              isThroughStreet(1, intersection),
                                              intersection[1]);
-                util::Log(logDEBUG) << "Disabled U-Turn on a freeway";
                 intersection[0].entry_allowed = false; // UTURN on the freeway
             }
             else if (exiting_motorways == 2)
@@ -334,8 +329,6 @@ Intersection MotorwayHandler::fromMotorway(const EdgeID via_eid, Intersection in
             }
             else
             {
-                util::Log(logDEBUG) << "Found motorway junction with more than "
-                                       "2 exiting motorways or additional ramps";
                 return fallback(std::move(intersection));
             }
         } // done for more than one highway exit
@@ -488,9 +481,7 @@ Intersection MotorwayHandler::fromRamp(const EdgeID via_eid, Intersection inters
         }
     }
     else
-    { // FALLBACK, hopefully this should never been reached
-        util::Log(logDEBUG) << "Reached fallback on motorway ramp with " << intersection.size()
-                            << " roads and " << countValid(intersection) << " valid turns.";
+    {
         return fallback(std::move(intersection));
     }
     return intersection;
