+++ src/extractor/guidance/node_based_graph_walker.cpp
@@ -1,5 +1,8 @@
 #include "extractor/guidance/node_based_graph_walker.hpp"
 #include "util/coordinate_calculation.hpp"
+#include <utility>
+
+using osrm::util::angularDeviation;
 
 namespace osrm
 {
