+++ src/extractor/guidance/turn_handler.cpp
@@ -1,8 +1,8 @@
 #include "extractor/guidance/turn_handler.hpp"
 #include "extractor/guidance/constants.hpp"
-#include "extractor/guidance/toolkit.hpp"
 
-#include "util/guidance/toolkit.hpp"
+#include "util/bearing.hpp"
+#include "util/guidance/name_announcements.hpp"
 
 #include <algorithm>
 #include <limits>
@@ -11,8 +11,8 @@
 #include <boost/assert.hpp>
 
 using EdgeData = osrm::util::NodeBasedDynamicGraph::EdgeData;
-using osrm::util::guidance::getTurnDirection;
-using osrm::util::guidance::angularDeviation;
+using osrm::extractor::guidance::getTurnDirection;
+using osrm::util::angularDeviation;
 
 namespace osrm
 {
