+++ src/extractor/guidance/turn_lane_data.cpp
@@ -5,7 +5,6 @@
 
 #include <algorithm>
 #include <cstddef>
-#include <string>
 #include <unordered_map>
 #include <utility>
