+++ src/extractor/guidance/turn_lane_matcher.cpp
@@ -1,11 +1,12 @@
 #include "extractor/guidance/turn_lane_matcher.hpp"
-#include "extractor/guidance/toolkit.hpp"
-#include "util/guidance/toolkit.hpp"
+#include "util/bearing.hpp"
 
 #include <boost/assert.hpp>
 #include <boost/numeric/conversion/cast.hpp>
 
 #include <functional>
+
+using osrm::util::angularDeviation;
 
 namespace osrm
 {
@@ -53,8 +54,8 @@ DirectionModifier::Enum getMatchingModifier(const TurnLaneType::Mask tag)
 // check whether a match of a given tag and a turn instruction can be seen as valid
 bool isValidMatch(const TurnLaneType::Mask tag, const TurnInstruction instruction)
 {
-    using util::guidance::hasLeftModifier;
-    using util::guidance::hasRightModifier;
+    using extractor::guidance::hasLeftModifier;
+    using extractor::guidance::hasRightModifier;
     const auto isMirroredModifier = [](const TurnInstruction instruction) {
         return instruction.type == TurnType::Merge;
     };
@@ -200,12 +201,13 @@ Intersection triviallyMatchLanesToTurns(Intersection intersection,
                                         const LaneDataVector &lane_data,
                                         const util::NodeBasedDynamicGraph &node_based_graph,
                                         const LaneDescriptionID lane_string_id,
-                                        LaneDataIdMap &lane_data_to_id)
+                                        util::guidance::LaneDataIdMap &lane_data_to_id)
 {
     std::size_t road_index = 1, lane = 0;
 
     const auto matchRoad = [&](ConnectedRoad &road, const TurnLaneData &data) {
-        LaneTupleIdPair key{{LaneID(data.to - data.from + 1), data.from}, lane_string_id};
+        util::guidance::LaneTupleIdPair key{{LaneID(data.to - data.from + 1), data.from},
+                                            lane_string_id};
 
         auto lane_data_id = boost::numeric_cast<LaneDataID>(lane_data_to_id.size());
         const auto it = lane_data_to_id.find(key);
