+++ src/util/coordinate_calculation.cpp
@@ -315,6 +315,49 @@ bool isCCW(const Coordinate first_coordinate,
     return signedArea(first_coordinate, second_coordinate, third_coordinate) > 0;
 }
 
+std::pair<util::Coordinate, util::Coordinate>
+leastSquareRegression(const std::vector<util::Coordinate> &coordinates)
+{
+    BOOST_ASSERT(coordinates.size() >= 2);
+    double sum_lon = 0, sum_lat = 0, sum_lon_lat = 0, sum_lon_lon = 0;
+    double min_lon = static_cast<double>(toFloating(coordinates.front().lon));
+    double max_lon = static_cast<double>(toFloating(coordinates.front().lon));
+    for (const auto coord : coordinates)
+    {
+        min_lon = std::min(min_lon, static_cast<double>(toFloating(coord.lon)));
+        max_lon = std::max(max_lon, static_cast<double>(toFloating(coord.lon)));
+        sum_lon += static_cast<double>(toFloating(coord.lon));
+        sum_lon_lon +=
+            static_cast<double>(toFloating(coord.lon)) * static_cast<double>(toFloating(coord.lon));
+        sum_lat += static_cast<double>(toFloating(coord.lat));
+        sum_lon_lat +=
+            static_cast<double>(toFloating(coord.lon)) * static_cast<double>(toFloating(coord.lat));
+    }
+
+    const auto dividend = coordinates.size() * sum_lon_lat - sum_lon * sum_lat;
+    const auto divisor = coordinates.size() * sum_lon_lon - sum_lon * sum_lon;
+    if (std::abs(divisor) < std::numeric_limits<double>::epsilon())
+        return std::make_pair(coordinates.front(), coordinates.back());
+
+    // slope of the regression line
+    const auto slope = dividend / divisor;
+    const auto intercept = (sum_lat - slope * sum_lon) / coordinates.size();
+
+    const auto GetLatAtLon = [intercept,
+                              slope](const util::FloatLongitude longitude) -> util::FloatLatitude {
+        return {intercept + slope * static_cast<double>((longitude))};
+    };
+
+    const util::Coordinate regression_first = {
+        toFixed(util::FloatLongitude{min_lon - 1}),
+        toFixed(util::FloatLatitude(GetLatAtLon(util::FloatLongitude{min_lon - 1})))};
+    const util::Coordinate regression_end = {
+        toFixed(util::FloatLongitude{max_lon + 1}),
+        toFixed(util::FloatLatitude(GetLatAtLon(util::FloatLongitude{max_lon + 1})))};
+
+    return {regression_first, regression_end};
+}
+
 } // ns coordinate_calculation
 } // ns util
 } // ns osrm
